requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.8
django-cleanup==9.0.0
djangorestframework==3.16.1
flatbuffers==25.9.23
fonttools==4.60.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
jax==0.6.2
jaxlib==0.6.2
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
llvmlite==0.45.1
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==1.26.4
onnxruntime==1.23.2
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==4.25.8
pycparser==2.23
PyMatting==1.1.14
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
referencing==0.37.0
rembg==2.0.68
requests==2.32.5
rpds-py==0.29.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.3
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
